services/appointments-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
python-multipart
passlib
python-jose
cryptography
email-validator
pydantic[email]